remove init

+++ src/main/java/main/com/skillbox/ru/developerspublics/service/PostService.java
@@ -46,7 +46,6 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.server.ResponseStatusException;
 
 
 @Service
@@ -83,28 +82,8 @@ public class PostService {
   }
 
 
-  private Post getInitPostById(int id) {
-    Post post = getPostById(id).orElseThrow(
-        () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Post id=" + id + " not found"));
-    initPost(post);
-    return post;
-  }
-
-
   private Optional<Post> getPostById(int id) {
     return postsRepository.findById(id);
-  }
-
-
-  private List<Post> getInitActivePosts() {
-    List<Post> activePosts = new ArrayList<>();
-    for (Post post : postsRepository.findByIsActiveAndModerationStatusAndTimeBefore(
-        1, ModerationStatuses.ACCEPTED.toString(), new Date(System.currentTimeMillis()),
-        null)) {
-      initPost(post);
-      activePosts.add(post);
-    }
-    return activePosts;
   }
 
 
@@ -154,9 +133,6 @@ public class PostService {
         break;
     }
 
-    if (posts.size() != 0) {
-      initPosts(posts);
-    }
     return posts;
   }
 
@@ -173,7 +149,7 @@ public class PostService {
 
 
   private List<Post> getActivePostsByQuery(String query, int offset, int limit) {
-    List<Post> posts = postsRepository
+    return postsRepository
         .findActivePostsByQuery(
             1,
             ModerationStatuses.ACCEPTED.toString(),
@@ -181,10 +157,7 @@ public class PostService {
             "%" + query + "%",
             PageRequest.of(offset / limit, limit, Sort.by("time"))
         );
-    initPosts(posts);
-    return posts;
   }
-
 
   @SneakyThrows
   private List<Post> getActivePostByDate(String date) {
@@ -201,16 +174,11 @@ public class PostService {
     thisDay.roll(Calendar.DATE, -1);
     Date dayBefore = thisDay.getTime();
 
-    List<Post> posts = postsRepository.findByIsActiveAndModerationStatusAndTimeAfterAndTimeBefore(
+    return postsRepository.findByIsActiveAndModerationStatusAndTimeAfterAndTimeBefore(
         1,
         ModerationStatuses.ACCEPTED.toString(),
         dayBefore,
         dayAfter);
-
-    for (Post post : posts) {
-      initPost(post);
-    }
-    return posts;
   }
 
 
@@ -248,18 +216,30 @@ public class PostService {
 
   @Transactional
   public void deletePost(Post post) {
-    initPost(post);
     Post postDB = postsRepository.findByTitle(post.getTitle());
       if (postDB != null) {
-          postsRepository.delete(postDB);
+        // delete tag2post
+        if (post.getTagToPosts() != null) {
+          for (TagToPost tagToPost : post.getTagToPosts()) {
+            Tag tag = tagService.findTagById(tagToPost.getTagId());
+            tagToPostService.deleteTagToPost(tagToPost);
+            // delete tags without posts
+            if (tagToPostService.getTagToPostsByTagId(tag.getId()).size() == 0) {
+              tagService.deleteTag(tag);
+            }
+          }
+        }
+        // delete post comments
+        if (post.getPostComments() != null) {
+          post.getPostComments().forEach(postCommentService::deletePostComment);
+        }
+        // delete post votes
+        if (post.getPostVotes() != null) {
+          post.getPostVotes().forEach(postVoteService::delete);
+        }
+        // delete post
+        postsRepository.delete(postDB);
       }
-    for (TagToPost tagToPost : post.getTagToPosts()) {
-      Tag tag = tagService.findTagById(tagToPost.getTagId());
-      tagToPostService.deleteTagToPost(tagToPost);
-      if (tagToPostService.getTagToPostsByTagId(tag.getId()).size() == 0) {
-        tagService.deleteTag(tag);
-      }
-    }
   }
 
 
@@ -298,17 +278,12 @@ public class PostService {
     } catch (Exception e) {
       return false;
     }
-
     return true;
   }
 
 
   private int countActivePosts() {
-    return postsRepository.findByIsActiveAndModerationStatusAndTimeBefore(
-        1,
-        ModerationStatuses.ACCEPTED.toString(),
-        new Date(System.currentTimeMillis()),
-        null).size();
+    return postsRepository.countActivePosts(new Date(System.currentTimeMillis()));
   }
 
 
@@ -390,27 +365,6 @@ public class PostService {
   }
 
 
-  @SneakyThrows
-  private void initPost(Post post) {
-    if (post.getModeratorId() != null) {
-      post.setModeratorPost(userService.findUserById(post.getModeratorId()));
-    }
-
-    post.setUserPost(userService.findUserById(post.getUserId()));
-    post.setPostVotes(postVoteService.findPostVotesByPostId(post.getId()));
-    post.setPostComments(postCommentService.getPostCommentsByPostId(post.getId()));
-    post.setTagToPosts(tagToPostService.getTagToPostsByPostId(post.getId()));
-  }
-
-
-  private void initPosts(List<Post> posts) {
-    for (Post post : posts) {
-      initPost(post);
-    }
-  }
-
-
-  @SneakyThrows
   private List<Post> sortPostList(List<Post> posts, int offset, int limit) {
     return posts.stream()
         .sorted(Comparator.comparing(Post::getTime).reversed())
@@ -420,7 +374,6 @@ public class PostService {
   }
 
 
-  @SneakyThrows
   private List<PostResponse> responsePosts(List<Post> posts) {
     List<PostResponse> list = new ArrayList<>();
     //приводим к нужному виду
@@ -441,7 +394,6 @@ public class PostService {
   }
 
 
-  @SneakyThrows
   public PostByIdResponse postByIdToJSON(Post post) {
     return new PostByIdResponse(
         post.getId(),
@@ -571,15 +523,19 @@ public class PostService {
   public ResponseEntity<?> getApiPostId(int id) {
     //ищем нужный пост по id
     //если пост не найден - возвращаем 404
-    Post post = getInitPostById(id);
+    Optional<Post> optional = getPostById(id);
+    if (optional.isEmpty()) {
+      return ResponseEntity.status(404).body(null);
+    }
+    Post post = optional.get();
 
     //При успешном запросе увеличиваем количество просмотров поста на 1, кроме случаев:
     // - Если модератор авторизован, то не считаем его просмотры вообще
     // - Если автор авторизован, то не считаем просмотры своих же публикаций
     Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-      if (authentication == null) {
-          return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
-      }
+    if (authentication == null) {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+    }
     User user = userService.findUserByLogin(authentication.getName());
 
     //смотрит не аноним
@@ -609,7 +565,6 @@ public class PostService {
   }
 
 
-  @SneakyThrows
   public ResponseEntity<?> getApiPostByDate(int offset, int limit, String date) {
     //получаем список постов за дату
     List<Post> posts = getActivePostByDate(date);
@@ -632,7 +587,6 @@ public class PostService {
       Optional<Post> optionalPost = getPostById(tagToPost.getPostId());
       if (optionalPost.isPresent()) {
         Post post = optionalPost.get();
-        initPost(post);
         //проверяем на активность
         if (isPostActive(post)) {
           //и запоминаем
@@ -658,6 +612,7 @@ public class PostService {
       }
     //выдергиваем из контекста пользователя
     User user = userService.findUserByLogin(authentication.getName());
+    System.out.println(user);
       if (user == null || user.getIsModerator() != 1) {
           return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
       }
@@ -670,7 +625,6 @@ public class PostService {
       //NEW - выводим все
       //ACCEPTED-DECLINED -> выводим только для текущего пользователя
       if (post.getIsActive() == 1) {
-        initPost(post);
           if (status.equals(ModerationStatuses.NEW.toString())) {
               posts.add(post);
           } else if (post.getModeratorId() != null && post.getModeratorId() == user.getId()) {
@@ -693,7 +647,6 @@ public class PostService {
     List<Post> posts = new ArrayList<>();
     for (Post post : findPostsByUserId(
         userService.findUserByLogin(authentication.getName()).getId())) {
-      initPost(post);
       //в зависимости от статуса добавляем нужные
       switch (status) {
         case "inactive":
@@ -874,7 +827,7 @@ public class PostService {
     int viewsCount = 0;
     //timestamp of first publication
     long firstPublication = System.currentTimeMillis() / 1000;
-    List<Post> activePosts = getInitActivePosts();
+    List<Post> activePosts = findActivePosts();
 
     //перебираем все активные посты
     for (Post post : activePosts) {
